+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Release Notes
 
-## Unreleased 
+## Unreleased
 
 * Fixed a bug where Document Capture would occasionally fail
+* Fixed a bug where Document Verification would get stuck on the processing screen
 
 ## 10.1.0